backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
pydantic==2.5.0
finnhub-python==2.4.19
alpha-vantage==2.3.1
pandas-market-calendars==4.1.4
beautifulsoup4==4.12.2
vaderSentiment==3.3.2
praw==7.7.1
aiohttp==3.9.1